pyproject.toml
[project]
name = "northwind-powerbi-to-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=46.0.3",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "snowflake-connector-python>=4.1.1",
    "streamlit>=1.52.1",
]
